重试retryCount由short改为int

+++ joyqueue-server/joyqueue-retry/joyqueue-retry-db/src/main/java/org/joyqueue/server/retry/db/DBMessageRetry.java
@@ -164,7 +164,7 @@ public class DBMessageRetry implements MessageRetry<Long> {
 
             consumeRetry.setExpireTime(getExpireTime(retryPolicy, SystemClock.now()));
             consumeRetry.setRetryTime(getRetryTime(retryPolicy, SystemClock.now(), 1));
-            consumeRetry.setRetryCount((short) 0);
+            consumeRetry.setRetryCount(0);
             consumeRetry.setData(retryMessageModel.getBrokerMessage());
             consumeRetry.setException(retryMessageModel.getException());
             consumeRetry.setCreateTime(SystemClock.now());
